initial Teardrop commit
twilio.go is the primary interface between TD and Twilio, the SMS
gateway.

items.go defines what objects to release, who to release them to, and
how long after the trigger moment to release them. It is pretty heavy on
Google Drive

teardrop.go will be the main powerhouse, but only has loading configs
and a very basic Run.